OJ-1184 pull out all types
