--- lambdas/src/common/config/cri-audit-config.ts
@@ -1,5 +1,0 @@
-export interface CriAuditConfig {
-    queueUrl: string;
-    auditEventNamePrefix: string;
-    issuer: string;
-}
