--- lambdas/src/common/config/jwt-verification-config.ts
@@ -1,4 +1,0 @@
-export interface JwtVerificationConfig {
-    publicSigningJwk: string;
-    jwtSigningAlgorithm: string;
-}
