--- lambdas/src/common/services/models/validation-result.ts
@@ -1,5 +1,0 @@
-export interface ValidationResult {
-    isValid: boolean;
-    errorMsg: string | null;
-    validatedObject?: any;
-}
