--- lambdas/src/common/services/session-item.ts
@@ -1,7 +1,0 @@
-export interface SessionItem {
-    sessionId: string;
-    clientId: string;
-    clientSessionId: string;
-    authorizationCode: string;
-    authorizationCodeExpiryDate: number;
-}
