+++ lambdas/src/handlers/session-handler.ts
@@ -4,15 +4,15 @@ import { Metrics, MetricUnits } from "@aws-lambda-powertools/metrics";
 import { Tracer } from "@aws-lambda-powertools/tracer";
 import { Logger } from "@aws-lambda-powertools/logger";
 import { ConfigService } from "../common/config/config-service";
-import { ClientConfigKey, CommonConfigKey } from "../common/config/config-keys";
+import { ClientConfigKey, CommonConfigKey } from "../types/config-keys";
 import { SessionService } from '../services/session-service';
 import { JweDecrypter } from "../services/security/jwe-decrypter";
 import { PersonIdentityService } from "../services/person-identity-service";
-import { PersonIdentity } from "../common/services/models/person-identity";
+import { PersonIdentity } from "../types/person-identity";
 import { SessionRequestValidatorFactory } from "../services/session-request-validator";
 import { AuditService } from "../common/services/audit-service";
-import { AuditEventType } from "../common/services/models/audit-event";
-import { SessionRequestSummary } from "../services/models/session-request-summary";
+import { AuditEventType } from "../types/audit-event";
+import { SessionRequestSummary } from "../types/session-request-summary";
 import { JWTPayload } from "jose";
 import { DynamoDBDocument } from "@aws-sdk/lib-dynamodb";
 import { SSMClient } from "@aws-sdk/client-ssm";
