--- lambdas/src/services/models/session-request-summary.ts
@@ -1,9 +1,0 @@
-export interface SessionRequestSummary {
-    clientId: string;
-    redirectUri: string;
-    subject: string;
-    persistentSessionId: string;
-    clientSessionId: string;
-    clientIpAddress: string | null;
-    state: string;
-}
