--- lambdas/src/services/session-request-validation-config.ts
@@ -1,5 +1,0 @@
-export interface SessionRequestValidationConfig {
-    expectedJwtRedirectUri: string;
-    expectedJwtIssuer: string;
-    expectedJwtAudience: string;
-}
