+++ lambdas/src/types/validation-result.ts
@@ -1,4 +1,5 @@
-export type ValidationResult = {
+export interface ValidationResult {
     isValid: boolean;
     errorMsg: string | null;
-};
+    validatedObject?: any;
+}
